updated python package with cli bin
