+++ .gitignore
@@ -1,2 +1,2 @@
 dist/*
-src/jetson_fan_ctl.egg-info/*
+src/jetsonfanctl.egg-info/*
