+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = jetson-fan-ctl
+name = jetsonfanctl
 version = 0.1.0
 author = Pyrestone
 author_email = author@example.com
@@ -17,6 +17,10 @@ package_dir =
     = src
 packages = find:
 python_requires = >=3.6
+
+[options.entry_points]
+console_scripts=
+  jetsonfanctl=jetsonfanctl.command_line:main
 
 [options.packages.find]
 where = src
