+++ src/jetsonfanctl/command_line.py
@@ -1,0 +1,4 @@
+from . import jetsonfanctl
+
+def main():
+  jetsonfanctl.main()
